checked the structure and everything is working now
