+++ middleware/authMiddleware.js
@@ -1,3 +1,6 @@
+const dotenv = require('dotenv');
+dotenv.config();
+
 const isAuthenticated = (req, res, next) => {
     if (req.isAuthenticated()) {
         return next();
@@ -5,4 +8,12 @@ const isAuthenticated = (req, res, next) => {
     res.status(401).json({ error: 'Unauthorized' });
 };
 
-module.exports = { isAuthenticated };
+const isAdmin = (req, res, next) => {
+    const adminIds = process.env.ADMIN_GITHUB_IDS.split(',');
+    if (req.user && adminIds.includes(req.user.username)) {
+        return next();
+    }
+    res.status(403).json({ error: 'Not authorized. Admin access required.' });
+};
+
+module.exports = { isAuthenticated, isAdmin };
